feat: 닉네임설정 절차지향적코드
